Hice el recorrido por niveles con una cola

+++ Aeropuertos.cpp
@@ -3,7 +3,7 @@
 
 //Constructor de la clase.
 //POST: Construye la clase con los datos proporcionados.
-Aeropuertos :: Aeropuertos(string nombre, string ciudad, string pais, int superficie, int cantidad_terminales, int destinos_nacionales, int destinos_internacionales){
+Aeropuertos :: Aeropuertos(string nombre, string ciudad, string pais, double superficie, int cantidad_terminales, int destinos_nacionales, int destinos_internacionales){
     this->nombre = nombre;
     this->ciudad = ciudad;
     this->pais = pais;
@@ -29,7 +29,7 @@ string Aeropuertos :: get_pais(){
 }
 
 //POST: Devuelve la superficie en hilometros cuadrados del aeropuerto.
-int Aeropuertos :: get_superficie(){
+double Aeropuertos :: get_superficie(){
     return this->superficie;
 }
